+++ lib/screen_record_plus.dart
@@ -1,6 +1,5 @@
 library screen_record_plus;
 
 export 'src/exporter.dart';
-export 'src/frame.dart';
 export 'src/screen_recorder.dart';
 export 'src/native_screen_recorder.dart';